gotowe

+++ src/main/java/entities/Movable.java
@@ -4,17 +4,19 @@ import javafx.application.Platform;
 import world.Position;
 
 public abstract class Movable extends Element {
-    protected int speed;
+    protected double speed;
 
     protected Position target;
     protected Position direction;
 
     public boolean move() {
         if (target.subtract(position).length() < speed) {
+            // Koniec ruchu
             return false;
-//            moved();
         }
+        // Rusz obiekt
         position = position.add(direction);
+        // rusz obraz
         Platform.runLater(() -> view.relocate(position.x, position.y));
         return true;
     }
@@ -22,6 +24,4 @@ public abstract class Movable extends Element {
     public Position calculateDirection() {
         return target.subtract(position).normalize(speed);
     }
-
-//    protected void moved() {}
 }
